Move Extension to Dependency Injection
