+++ source/Databricks/source/SqlStatementExecution.Tests/DatabricksSqlChunkResponseParserTests.cs
@@ -15,7 +15,7 @@
 using Energinet.DataHub.Core.Databricks.SqlStatementExecution.Internal;
 using Energinet.DataHub.Core.TestCommon.AutoFixture.Attributes;
 
-namespace Energinet.DataHub.Core.Databricks.SqlStatementExecutionTests;
+namespace Energinet.DataHub.Core.Databricks.SqlStatementExecution.Tests;
 
 public class DatabricksSqlChunkResponseParserTests
 {
