+++ source/Databricks/source/SqlStatementExecution.Tests/DatabricksSqlResponseParserTests.cs
@@ -21,7 +21,7 @@ using Moq;
 using Newtonsoft.Json.Linq;
 using Xunit.Categories;
 
-namespace Energinet.DataHub.Core.Databricks.SqlStatementExecutionTests;
+namespace Energinet.DataHub.Core.Databricks.SqlStatementExecution.Tests;
 
 [UnitTest]
 public class DatabricksSqlResponseParserTests
