+++ source/Databricks/source/SqlStatementExecution.Tests/SqlStatementExecution.Tests.csproj
@@ -6,8 +6,8 @@
     <Nullable>enable</Nullable>
 
     <IsPackable>false</IsPackable>
-    <AssemblyName>Energinet.DataHub.Core.Databricks.SqlStatementExecutionTests</AssemblyName>
-    <RootNamespace>Energinet.DataHub.Core.Databricks.SqlStatementExecutionTests</RootNamespace>
+    <AssemblyName>Energinet.DataHub.Core.Databricks.SqlStatementExecution.Tests</AssemblyName>
+    <RootNamespace>Energinet.DataHub.Core.Databricks.SqlStatementExecution.Tests</RootNamespace>
   </PropertyGroup>
 
   <ItemGroup>
